os/toolchains: Sleep before one of the architecture build branches
When the same job is built twice at the same instant, two instances
of Git are being run to update the pipeline repo refs.  This causes
failures when one of the jobs has its refs updated while it's still
updating them itself.

If new architectures are ever added, remember to add a different
sleep time to its branch here.