os/keyring: Add a job to automatically generate GPG keyrings
The intent here is to receive updated keys automatically from key
servers.  The job takes a list of key IDs to fetch and optionally a
pregenerated keyring to use as a starting point.  Whenever the job
is rerun, the parameters use their values from the previous run.  A
periodic build trigger will therefore produce a different keyring
only if any of the fetched keys have been updated.