os: Replace MANIFEST_REF with MANIFEST_TAG
Tags were always used anyway, and there were a few places that
expected a tag.  Now, with --verify-signature used unconditionally,
everything expects a tag.  Renaming the parameter drops any
ambiguity or implications that any ref can be used.  The only
exception is the /os/manifest job, which does take any type of ref,
then it creates a signed tag from it for the downstream jobs.