os/board/vm-matrix: Reorder VM image formats
As long as the VM images are still built in sequence(s) as opposed
to throwing them all at the scheduler and hoping for the best, this
will reach the point of being able to run "plume pre-release" a bit
sooner.  In particular, GCE, Azure, and the AMIs should be built
early.  It also leaves the QEMU and bare-metal-ish formats early in
the list.  Other than that, the list is alphabetized to make it
easier to find formats.  The ARM formats are just all alphabetized
with no concern for prioritizing them, since they are so few.