os: Clean up before adding prerelease jobs
There was a bit of reordering for alphabetical consistency, but a
big change is to drop the AWS bucket parameter.  Also, "AWS_DEV_"
was switched to "AWS_TEST_" in the parameter names.  The bucket
will be chosen by plume specs, and the two credentials differ by
being used for running kola tests and release commands.