os/manifest: Report effective changes built into official releases
This will hopefully be replaced by automated release notes one day.

Note that other repositories like mantle and jenkins-os can still
affect releases, but they are not included here because they are
not anchored to manifest tags.  (The mantle version used during the
OS build was built from master, not the commit in the manifest.)