os: Defer to the scripts repository for OS build commands
This supports branching the scripts' usage with releases.  Only the
primary OS/SDK jobs are affected here.  Jenkins sets up the whole
environment and securely initializes the workspace SDK, then the
appropriate script is run with the actual build commands.

A minor change to make this fit together is that the list of VM
formats to build is now given as a parameter.  This could be useful
for rebuilding a subset of images if something went wrong.