os/kola/dev-container: Use the kernel config that was just built
The original commands were taken from the documentation, which
expects to build modules for the host kernel.  The host kernel will
not always match the version being built and tested here, though,
so this takes the config from the binary package that was just
built.  That package has to be downloaded and extracted in tmpfs to
avoid filling the disk image.