os/kola/oci: fix file permissions
Currently the `oci` command-line tool spits out WARNING messages telling
the user to run a setup command to repair the file permissions. This
ends up amounting to changing the file mode of the config & RSA private
key to 0600. This PR pre-emptively changes the file mode as soon as they
are moved from the dynamic location created via credentials.