os/kola/dev-container: Use the master branch's kernel correctly
The emerge-gitclone command expects to work with release numbers to
determine where to find the kernel.  When building off master, the
major version number is the same as the latest alpha.

To work around this, test if the version string contains a "master"
substring, and check out the master branches' ebuilds if so.