os/kola/dev-container: Check out manifest commits in the repos
The emerge-gitclone script assumes it's running in a released
developer container image, so it doesn't work properly in some test
builds (e.g. when testing a kernel PR or building on an unreleased
SDK version).

This replaces the emerge-gitclone command with reading commits out
of the manifest used in the current Jenkins run.  That will
correctly read pull request refs etc.  It's also a workaround for
running out of space from the binary package cache by mounting over
/var/lib/portage.