os/update-linux: Update the alpha and beta branches
This generates additional overlay pull requests for the alpha and
beta branches, determined from the most recent version published in
the release documentation.

Test builds are run on master, alpha, then beta in sequence so that
workspaces are reused (rather that duplicated to run in parallel)
to save disk space.  If disk space is not a concern, the "wait"
option on their build steps can be set to "false" to let them run
in parallel to get results faster.